now receive keyboard inputs; add boilerplate for all exception handling; interpret scan codes

+++ src/util.rs
@@ -1,4 +1,5 @@
-pub(crate) fn int_to_bytes(n: i32) -> &'static [u8] {
+
+pub(crate) fn int_to_bytes(n: i64) -> &'static [u8] {
     static mut BUFFER: [u8; 12] = [0; 12];
     let mut i = 11;
     let mut num = if n < 0 { -n } else { n };
